+++ unikmer/cmd/duplicated.go
@@ -1,0 +1,173 @@
+// Copyright © 2018 Wei Shen <shenwei356@gmail.com>
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package cmd
+
+import (
+	"bufio"
+	"encoding/binary"
+	"fmt"
+	"io"
+	"os"
+	"runtime"
+
+	"github.com/shenwei356/unikmer"
+	"github.com/spf13/cobra"
+	boom "github.com/tylertreat/BoomFilters"
+)
+
+// duplicatedCmd represents
+var duplicatedCmd = &cobra.Command{
+	Use:     "dup",
+	Aliases: []string{"duplicated"},
+	Short:   "find duplicated k-mers",
+	Long: `find duplicated k-mers (experimental)
+
+Attentions:
+  1. Inverse Bloom Filter is used to test k-mer duplication,
+     which may reports a false negative but can never reports a false positive.
+     Therefore, outputed k-mers are indeed duplicated, but it may misses some.
+
+`,
+	Run: func(cmd *cobra.Command, args []string) {
+		opt := getOptions(cmd)
+		runtime.GOMAXPROCS(opt.NumCPUs)
+
+		var err error
+
+		var files []string
+		infileList := getFlagString(cmd, "infile-list")
+		if infileList != "" {
+			files, err = getListFromFile(infileList)
+			checkError(err)
+		} else {
+			files = getFileList(args)
+		}
+
+		checkFiles(extDataFile, files...)
+
+		outFile := getFlagString(cmd, "out-prefix")
+		capacity := getFlagPositiveInt(cmd, "capacity")
+
+		ibf := boom.NewInverseBloomFilter(uint(capacity))
+
+		if !isStdout(outFile) {
+			outFile += extDataFile
+		}
+		outfh, gw, w, err := outStream(outFile, opt.Compress, opt.CompressionLevel)
+		checkError(err)
+		defer func() {
+			outfh.Flush()
+			if gw != nil {
+				gw.Close()
+			}
+			w.Close()
+		}()
+
+		var writer *unikmer.Writer
+
+		var infh *bufio.Reader
+		var r *os.File
+		var reader *unikmer.Reader
+		var kcode unikmer.KmerCode
+		var k int = -1
+		var canonical bool
+		var firstFile = true
+		var n int64
+		var flag int
+		var nfiles = len(files)
+		buf := make([]byte, 8)
+		be := binary.BigEndian
+		for i, file := range files {
+			if !firstFile && file == files[0] {
+				continue
+			}
+
+			if opt.Verbose {
+				log.Infof("processing file (%d/%d): %s", i+1, nfiles, file)
+			}
+
+			flag = func() int {
+				infh, r, _, err = inStream(file)
+				checkError(err)
+				defer r.Close()
+
+				reader, err = unikmer.NewReader(infh)
+				checkError(err)
+
+				if k == -1 {
+					k = reader.K
+					canonical = reader.Flag&unikmer.UNIK_CANONICAL > 0
+
+					var mode uint32
+					if opt.Compact {
+						mode |= unikmer.UNIK_COMPACT
+					}
+					if canonical {
+						mode |= unikmer.UNIK_CANONICAL
+					}
+					writer, err = unikmer.NewWriter(outfh, k, mode)
+					checkError(err)
+				} else if k != reader.K {
+					checkError(fmt.Errorf("K (%d) of binary file '%s' not equal to previous K (%d)", reader.K, file, k))
+				} else if (reader.Flag&unikmer.UNIK_CANONICAL > 0) != canonical {
+					checkError(fmt.Errorf(`'canonical' flags not consistent, please check with "unikmer stats"`))
+				}
+
+				for {
+					kcode, err = reader.Read()
+					if err != nil {
+						if err == io.EOF {
+							break
+						}
+						checkError(err)
+					}
+
+					// new kmers
+					be.PutUint64(buf, kcode.Code)
+					if ibf.TestAndAdd(buf) {
+						n++
+						writer.Write(kcode) // not need to check err
+					}
+				}
+
+				return flagContinue
+			}()
+
+			if flag == flagReturn {
+				return
+			} else if flag == flagBreak {
+				break
+			}
+		}
+
+		checkError(writer.Flush())
+		if opt.Verbose {
+			log.Infof("%d k-mers saved", n)
+		}
+	},
+}
+
+func init() {
+	RootCmd.AddCommand(duplicatedCmd)
+
+	duplicatedCmd.Flags().StringP("out-prefix", "o", "-", `out file prefix ("-" for stdout)`)
+	duplicatedCmd.Flags().IntP("capacity", "", 100000000, `capacity of Inverse Bloom Filter`)
+}